Separated middleware into a package.
