+++ middleware/database_middleware.py
@@ -1,0 +1,47 @@
+from functools import wraps
+from typing import Callable
+
+import discord
+
+from database.models.analytics_model import Analytics
+from database.models.server_model import Server
+
+
+def ensure_server_document(func: Callable) -> Callable:
+    @wraps(func)
+    async def wrapper(self, interaction: discord.Interaction, *args, **kwargs) -> Callable | None:
+        server_id = interaction.guild.id
+        server = await Server.get(server_id)
+
+        if not server:
+            server = Server(id=server_id)
+            await server.create()
+
+        return await func(self, interaction, *args, **kwargs)
+
+    return wrapper
+
+
+def track_analytics(func: Callable) -> Callable:
+    @wraps(func)
+    async def wrapper(self, interaction: discord.Interaction, *args, **kwargs) -> Callable | None:
+        analytics = await Analytics.find_all().to_list()
+
+        if not analytics:
+            analytics = Analytics()
+            await analytics.create()
+        else:
+            analytics = analytics[0]
+
+        if interaction.user.id not in analytics.distinct_users_total:
+            analytics.distinct_users_total.append(interaction.user.id)
+
+        if interaction.user.id not in analytics.distinct_users_today:
+            analytics.distinct_users_today.append(interaction.user.id)
+
+        analytics.command_count_today += 1
+        await analytics.save()
+
+        return await func(self, interaction, *args, **kwargs)
+
+    return wrapper
